refactor: restructured the project to separate frontend and backend.
- Moved the server.js file to a new dedicated backend folder.
- Created the backend structure (Controllers, Middleware, and Routes).
- Installed the necessary dependencies for the backend functionality.
- Removed entries from the frontend package.json related to unused backend files.
