+++ backend/controllers/transactionController.js
@@ -1,0 +1,7 @@
+const db = require("../db/transactions.json");
+
+const listTransactions = (req, res) => {
+  return res.json(db);
+};
+
+module.exports = { listTransactions };
