+++ backend/routes/authRoutes.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const { authenticate } = require("../controllers/authControllers");
+
+const router = express.Router();
+
+router.post("/", authenticate);
+
+module.exports = router;
