+++ frontend/src/styles/global.ts
