+++ frontend/src/styles/variables.ts
