+++ frontend/src/vite-env.d.ts
