+++ frontend/tsconfig.json
@@ -20,6 +20,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
+  "include": ["src", "src/main.tsx", "vite-env.d.ts", "../src/App.tsx"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
